daemonize websocket thread (kill when main exits)
from https://docs.python.org/2/library/threading.html:

```
A thread can be flagged as a “daemon thread”. The significance of
this flag is that the entire Python program exits when only daemon
threads are left. The initial value is inherited from the creating
thread. The flag can be set through the daemon property.
```