Accept kwargs in buy and sell
Previously buy and sell accepted a single dict of arbitrary parameters.
That worked, but it wasn't very pythonic.

This commit changes them both to accept arbitrary kwargs. The
functionality is the same, but now it looks like python.

I have also taken the opportunity to format the lines I altered in
accordance with PEP-8.