Use explicit classes when calling super()
In python3, we can just say `super().__init__()`. For python2
compatibility, we need to provide two parameters to `super`:
the inheriting class, and the bound object.

In 2655428, we introduced calls to `super` specifying `self.__class__`
for the first parameter. This is a problem (at least in python3) because
that results in the inheriting class itself used for binding. Thus when
we call `__init__`, it is not the `super`'s `__init__` that is called.

This change-set fixes that problem by explicitly specifying the first
parameter of `super` calls. This works in both python3 and python2.