Clean up to fix a thead joining issue
There were a couple of places where an OrderBook() calls `close()` on itself.
Since `OrderBook()` is derived from  `WebsocketClient()`, ultimately, the call
resolves to `WebsocketClient.close()`.  `WebsocketClient.close()` tries to call
`join()` on the thread it is currently running on which is not allowed.

To solve these to issues, WebsocketClient() was changed such that it closes the
web socket on the thread it is created when is calls _disconnect().

An error attribute was also added.  If on_error is called, the default action is
now to set the thread to stop and set the error attribute value to be the exception
that occured.  The user can then decide what to when the WebsocketClient() returns
with an error being set.  This behavior can be overridden by a subclass if desired.

The order book now attempts to recover the book after a sequence gap by resetting
the book and grabbing a new snapshot, rather than trying to close() and start().
Previously, it was creating a new thread from the old thread which would leave
dangling threads.

The order book no longer has an on_error() method (since it did basically the
same thing as toe old book recovery was doing) and instead relies on the default
action described above.