Fix creation of stop orders
A 'stop' isn't an actual order type but is actually a special flavour of
a limit order. The 'stop' and 'stop_price' params need to be set.

There are 2 stop types:
1. loss (triggers at or below the stop price) -> this has to be a sell
order to be valid (the Coinbase Pro will respond with an error
otherwise)
2. entry (trigger at or above the stop price) -> has to be a buy

Modified existing and added new test.