Request limit in paginated message
When you invoke _send_paginated_message, it will reach the Coinbase Pro API's rate limit, when the results have too many pages. This patch adds a sleep interval after each page that will work for both public and private endpoints, regardless of the number of pages.

This sleep will inevitably slow down paginated calls, but that's required as the number of pages requested grows without bound, per the Coinbase Pro API's spec's rate limits :
https://docs.pro.coinbase.com/#rate-limits

An additional improvement would be for the private endpoints to send the new sleep_interval parameter at .2 (as opposed to the default .34).